requirements.txt
fastapi
uvicorn
google-cloud-documentai
google-cloud-pubsub
pandas
pypdf
rapidfuzz
openpyxl
python-multipart
dotenv
fuzzywuzzy
python-Levenshtein